remove unnecessary component
